[fix/faq_entity](#35) Fix FAQ, change FAQRepositoryTest to FAQTest
src/main/java/org/rubatophil/www/api/domain
src/test/java/org/rubatophil/www/api/domain
src/test/java/org/rubatophil/www/api/repository
